+++ packages/antd/src/AntDValueSelector.tsx
@@ -2,8 +2,7 @@ import { Select } from 'antd';
 import type { ComponentPropsWithoutRef } from 'react';
 import * as React from 'react';
 import type { VersatileSelectorProps } from 'react-querybuilder';
-import { useValueSelector } from 'react-querybuilder';
-import { toOptions } from './utils';
+import { joinWith, useValueSelector } from 'react-querybuilder';
 
 export type AntDValueSelectorProps = VersatileSelectorProps &
   Omit<ComponentPropsWithoutRef<typeof Select>, 'onChange' | 'defaultValue'>;
@@ -31,21 +30,47 @@ export const AntDValueSelector = ({
   schema: _schema,
   ...extraProps
 }: AntDValueSelectorProps): React.JSX.Element => {
-  const { onChange, val } = useValueSelector({ handleOnChange, listsAsArrays, multiple, value });
+  // Alternate onChange handler that doesn't use arrays even when `multiple` is true
+  const { onChange: onChangeNoArrays } = useValueSelector({
+    handleOnChange,
+    listsAsArrays: false,
+    multiple: false,
+    value,
+  });
+  const { onChange: onChangeNormal, val } = useValueSelector({
+    handleOnChange,
+    // This forces `val` to be an array if `multiple` is true,
+    // even if `listsAsArrays` is false
+    listsAsArrays: multiple || listsAsArrays,
+    multiple,
+    value,
+  });
 
-  const modeObj = multiple ? { mode: 'multiple' as const } : {};
+  const onChange = React.useCallback(
+    (v: string | string[]) => {
+      if (multiple && !listsAsArrays && Array.isArray(v)) {
+        // `multiple: true` means `v` is probably an array, but we don't want
+        // to send an array to `handleOnChange` when `listsAsArrays` is false
+        onChangeNoArrays(joinWith(v));
+      } else {
+        onChangeNormal(v);
+      }
+    },
+    [listsAsArrays, multiple, onChangeNoArrays, onChangeNormal]
+  );
 
   return (
-    <span title={title} className={className}>
-      <Select
-        {...modeObj}
-        popupMatchSelectWidth={false}
-        disabled={disabled}
-        value={val}
-        onChange={onChange}
-        {...extraProps}>
-        {toOptions(options)}
-      </Select>
-    </span>
+    <Select
+      {...(multiple ? { mode: 'multiple', allowClear: true } : {})}
+      title={title}
+      className={className}
+      popupMatchSelectWidth={false}
+      disabled={disabled}
+      value={val}
+      onChange={onChange}
+      optionFilterProp="label"
+      options={options}
+      {...extraProps}
+    />
   );
 };