+++ website/src/pages/demo/antd.tsx
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/consistent-type-imports */
+import '@ant-design/v5-patch-for-react-19';
 import BrowserOnly from '@docusaurus/BrowserOnly';
 import { useColorMode } from '@docusaurus/theme-common';
 import { AntDValueSelector, QueryBuilderAntD } from '@react-querybuilder/antd';