+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-N/A
+### Fixed
+
+- [#836] `AntDValueSelector` now filters on the `label` property and defines options with the `options` prop instead of a `children` array of `<Select.OptGroup>`/`<Select.Option>`.
+- [#836] Utilities that determine whether arrays are valid option lists (`isFullOptionArray`, `isFlexibleOptionArray`, `isFullOptionGroupArray`, `isFlexibleOptionGroupArray`) are now more strict: All leaf elements must conform to the operative type, not just the first element.
 
 ## [v8.1.0] - 2025-01-09
 
@@ -1816,6 +1819,7 @@ Maintenance release focused on converting to a monorepo with Vite driving the bu
 [#814]: https://github.com/react-querybuilder/react-querybuilder/pull/814
 [#820]: https://github.com/react-querybuilder/react-querybuilder/pull/820
 [#834]: https://github.com/react-querybuilder/react-querybuilder/pull/834
+[#836]: https://github.com/react-querybuilder/react-querybuilder/pull/836
 
 <!-- #endregion -->